Improved folder structure by renaming 'container' to 'layout' for layout components
