+++ frontend/src/App.tsx
@@ -10,8 +10,8 @@ import * as Jsondata from './data/data.json';
 import Banner from './components/Banner';
 import BikeStation from './components/BikeStation';
 import BuyPass from './components/BuyPass';
-import Contact from './container/Contact';
-import Navigation from './container/Navigation';
+import Contact from './components/layout/Contact';
+import Navigation from './components/layout/Navigation';
 import Tickets from './components/Tickets';
 import TicketsPrice from './components/TicketsPrice';
 import LoginForm from './components/LoginForm';
